python/core/requirements.txt
# --------- pytorch --------- #
torch>=1.0.0

# --------- transformers --------- #
transformers>=4.0.0

# --------- train --------- #
pytorch-lightning>=1.5.10
torchmetrics>=0.7.0
datasets>=1.0.0

# --------- hydra --------- #
hydra-core>=1.1.0
hydra-colorlog>=1.1.0

# --------- loggers --------- #
wandb
# neptune-client
# mlflow
# comet-ml
# tensorboard

# --------- others --------- #
pyrootutils     # standardizing the project root setup
pre-commit      # hooks for applying linters on commit
rich            # beautiful text formatting in terminal
pytest          # tests
sh              # for running bash commands in some tests
